Add primary flag to packages
Addresses the issue where a product is available in multiple package sizes
but one of the sizes is considered the primary, or canonical, package that
is used to calculate stock (the other size(s) being based on the primary
package size).

For example, a vaccine is available in both 10- and 20-dose sizes but the
stock is calculated based on the 10-dose package. Any allocation of the
20-dose package would be treated as 2 10-dose packages.

Related forms are modified to disable the ability to set the values for
the non-primary package sizes, except for the quantity furnished field
on the after warehouse pickup form.